+++ changelog.md
@@ -3,6 +3,8 @@
 ## Upcoming
 * Added `heal_on_skull_kill` that defines how much health the wither regenerates when a skull kills an entity
   * 5/7/9/10 healing at I to IV. Was 10
+* Wither minion health is now configurable
+  * 10/12/14/15 health at I to IV. Was 15
 * Renamed `resistances_weaknesses` to `resistances` in Wither's json
 
 ## Alpha 4.1.1